requirements.txt
qq-botpy
pydub
fish-audio-sdk
pilk
request
markdown
imgkit
pymdown-extensions
Pygments
sanic